Change successMessage, adjust tests
